+++ AncientArcher/cpp/engine/AAControls.h
@@ -4,8 +4,8 @@ class AAControls
 {
 public:
   static AAControls* getInstance();
-  void mouseMovement(int xpos, int ypos);
-  void mouseScrollWheelMovement(int xpos, int ypos);
+  void mouseMovement(float xpos, float ypos);
+  void mouseScrollWheelMovement(float xpos, float ypos);
   void keyInput(AAKeyInput* keyInput);
 private:
 };
